% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_W}
\alias{generate_W}
\title{Generate Spatial Weight Matrix}
\usage{
generate_W(states)
}
\arguments{
\item{states}{A character vector of state abbreviations.}
}
\value{
A spatial weight matrix.
}
\description{
Generates a row-normalized spatial weight matrix based on state contiguity.
}
