\name{pcdid-package}
\alias{pcdid-package}
\alias{pcdid}
\docType{package}
\title{
\packageTitle{pcdid}
}
\description{
\packageDescription{pcdid}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{pcdid}

Maintainer: \packageMaintainer{pcdid}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = pcdid)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{pcdid}}
%\section{Documentation Index}{\packageIndices{pcdid}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
