% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdid.R
\name{pcdid}
\alias{pcdid}
\title{Principal Components Difference-in-Differences}
\usage{
pcdid(
  formula,
  index,
  data,
  alpha = FALSE,
  fproxy = NULL,
  stationary = FALSE,
  kmax = 10,
  nwlag = round(max(data[[index[2]]])^0.25)
)
}
\arguments{
\item{formula}{regression specification}

\item{index}{vector of length 2 indicating c(id, time)}

\item{data}{a data frame containing variables to be used}

\item{alpha}{whether to perform the alpha test}

\item{fproxy}{number of factors}

\item{stationary}{whether the factors are stationary}

\item{kmax}{maximum number of factors in growth ratio test}

\item{nwlag}{Newey-West lag for standard error correction}
}
\value{
A list of class \code{pcdid}, the output list includes element:

\describe{
 \item{mg}{mean-group estimate of the treatment effect}
 \item{alpha}{alpha test result}
 \item{treated}{regression result for treated units}
 \item{control}{regression result for control units}
}
}
\description{
Estimate treatment effect using the PCDID estimator.
}
\examples{
result <- pcdid(
  lncase ~ treated + treated_post + afdcben + unemp + empratio + mon_d2 + mon_d3 + mon_d4,
  index = c("state", "trend"),
  data = welfare,
  alpha = TRUE
)

result$mg

}
\author{
Xiaolei Wang \email{adamwang15@gmail.com}
}
