% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfdid.R
\name{sfdid}
\alias{sfdid}
\title{Spatial Factor Difference-in-Differences}
\usage{
sfdid(
  formula,
  index,
  data,
  W,
  npc_method = "BN",
  kmax_y = 5,
  kmax_c = 5,
  qml_iters = 15,
  tol_bhat = 1e-04,
  rho_hetero = 1
)
}
\arguments{
\item{formula}{regression specification: depvar ~ treatvar + didvar + indepvar}

\item{index}{vector of length 2 indicating c(id, time)}

\item{data}{a data frame containing variables to be used}

\item{W}{spatial weight matrix (N x N)}

\item{npc_method}{method for determining the number of factors: "BN" (Bai-Ng) or "ER" (Eigenvalue Ratio)}

\item{kmax_y}{max factors for L-step}

\item{kmax_c}{max factors for D-step}

\item{qml_iters}{max iterations for QML}

\item{tol_bhat}{tolerance for early stopping}

\item{rho_hetero}{spatial parameter heterogeneity setup (1 = full heterogeneous, 2, 3, 4 = homogeneous)}
}
\value{
A list of class \code{sfdid}
}
\description{
sfdid implements the Spatial Factor Difference-in-Differences estimator.
}
