% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{welfare}
\alias{welfare}
\title{Welfare caseloads data}
\format{
A data frame
\describe{
  \item{state}{TODO}
  \item{statenum}{TODO}
  \item{trend}{TODO}
  \item{treated}{TODO}
  \item{treated_post}{TODO}
  \item{lncase}{TODO}
  \item{afdcben}{TODO}
  \item{unemp}{TODO}
  \item{empratio}{TODO}
  \item{mon_d2}{TODO}
  \item{mon_d3}{TODO}
  \item{mon_d4}{TODO}
  \item{caseload}{TODO}
  \item{popn}{TODO}
  \item{empratio_raw}{TODO}
  \item{south}{TODO}
  \item{control}{TODO}
  \item{T0}{TODO}
}
}
\source{
Supplemental material, \url{https://doi.org/10.1080/07350015.2021.1914636}
}
\usage{
data(welfare)
}
\description{
A sample data to examine the effects of welfare waiver programs on welfare caseloads in the United States.
}
\references{
Chan, M. K., & Kwok, S. S. (2022). The PCDID approach: difference-in-differences when trends are potentially unparallel and stochastic. Journal of Business & Economic Statistics, 40(3), 1216-1233.
}
\keyword{datasets}
