% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdid-package.R
\docType{package}
\name{pcdid-package}
\alias{pcdid-package}
\title{pcdid: Principal Components Difference-in-Differences}
\description{
Implements the Principal Components Difference-in-Differences estimators as described in Chan, M. K., & Kwok, S. S. (2022) \doi{10.1080/07350015.2021.1914636}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamwang15/pcdid}
  \item Report bugs at \url{https://github.com/adamwang15/pcdid/issues}
}

}
\author{
\strong{Maintainer}: Xiaolei Wang \email{adamwang15@gmail.com} (\href{https://orcid.org/0009-0005-6192-9061}{ORCID})

Authors:
\itemize{
  \item Marc Chan \email{marc.chan@unimelb.edu.au} (\href{https://orcid.org/0000-0002-1010-7587}{ORCID})
}

}
\keyword{internal}
