% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdid-package.R
\docType{package}
\name{pcdid-package}
\alias{pcdid-package}
\title{pcdid: Principal Components Difference-in-Differences}
\description{
This package implements factor-augmented difference-in-differences (DID) estimation. It is useful in situations where the user suspects that trends may be unparallel and/or stochastic among control and treated units. The data structure is similar to that in a DID setup. The estimation method is regression-based and can be considered as an extension of conventional DID regressions. pcdid also implements a parallel trend alpha test (based on an interactive effects structure) and a recursive procedure that determines the number of factors automatically. For further details, please see Chan and Kwok (2016, 2020) who developed the pcdid approach and the alpha test.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamwang15/pcdid}
  \item Report bugs at \url{https://github.com/adamwang15/pcdid/issues}
}

}
\author{
\strong{Maintainer}: Xiaolei Wang \email{adamwang15@gmail.com} (\href{https://orcid.org/0009-0005-6192-9061}{ORCID})

Authors:
\itemize{
  \item Marc Chan \email{marc.chan@unimelb.edu.au} (\href{https://orcid.org/0000-0002-1010-7587}{ORCID})
}

}
\keyword{internal}
